@@ -1,6 +1,6 @@
 cmake_minimum_required(VERSION 3.12)
-project(MigangBotCPP)
-set(CMAKE_CXX_STANDARD 17)
+project(MigangBot)
+set(CMAKE_CXX_STANDARD 20)
 
 set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib)
 set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib)
