@@ -6,15 +6,25 @@
 #include <vector>
 #include <functional>
 #include <initializer_list>
+#include <filesystem>
+#include <fstream>
 
 #include "plugins/utility.h"
 #include "bot/api_bot.h"
 #include "utility"
 #include "event/event_handler.h"
 #include "event/event.h"
+#include "logger/logger.h"
 
 namespace white
 {
+
+using Config = YAML::Node;
+
+inline void RegisterAllMessage(const plugin_func &&func)
+{
+    EventHandler::GetInstance().RegisterCommand(ALLMSG, "", std::move(func));
+}
 
 inline void RegisterCommand(const int type, const std::initializer_list<std::string> &commands, const plugin_func &func)
 {
@@ -38,9 +48,55 @@ inline void RegisterRequest(const std::string &request_type, const std::string &
 class PluginInterface
 {
 public:
+    // 如果没有配置文件，就留空
+    PluginInterface(const std::string &config_file, const std::string &config_example = "")
+        : config_file_(config_file),
+        config_example_(config_example)
+    {}
+
     virtual void Register() = 0;
-    
+
+protected:
+    YAML::Node LoadConfig()
+    {
+        YAML::Node ret;
+        if(config_file_.empty())
+        {
+            LOG_WARN("未配置配置文件路径", config_file_);
+            return ret;
+        }
+        if(!std::string_view(config_file_).ends_with("yml"))
+        {
+            LOG_WARN("配置文件格式错误，应当为.yml");
+            return ret;
+        }
+        std::filesystem::path path;
+        if(config_file_[0] != '/')
+            path = kConfigDir / config_file_;
+        if(!std::filesystem::exists(path))
+        {
+            LOG_WARN("无法加载配置文件[{}]", config_file_);
+            LOG_INFO("尝试生成配置文件");
+            auto parent_path = path.parent_path();
+            if(!std::filesystem::create_directories(parent_path))
+            {
+                LOG_ERROR("创建配置文件[{}]失败", config_file_);
+                return ret;
+            }
+            std::fstream f(path, std::ios_base::out);
+            f << config_example_;
+            f.close();
+            LOG_INFO("已成功创建[{}]", config_file_);
+        }
+        ret = YAML::LoadFile(path);
+        return ret;
+    }
+
+private:
+    const std::string config_file_;
+    const std::string config_example_;
+
 };
 
-}
+} // namespace white
 #endif
